backend/requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.2.0
celery==5.3.6
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap4==2024.1
cryptography==41.0.3
defusedxml==0.7.1
Django==4.2.5
django-cors-headers==4.2.0
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
django-elasticsearch-dsl==8.0
django-filter==24.1
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.2
drf-yasg==1.21.7
elastic-transport==8.12.0
elasticsearch==8.12.1
elasticsearch-dsl==8.12.0
flower==2.0.1
frozenlist==1.4.0
gunicorn==21.2.0
humanize==4.9.0
idna==3.4
inflection==0.5.1
kombu==5.3.6
multidict==6.0.4
oauthlib==3.2.2
packaging==24.0
Pillow==10.0.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
tornado==6.4
twilio==8.8.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.2
